--- a/app/controllers/api/locale.php
+++ b/app/controllers/api/locale.php
@@ -30,16 +30,16 @@ App::get('/v1/locale')
         $currency = null;
 
         try {
-            $record = $geodb->country($ip);
-            $output['countryCode'] = $record->country->isoCode;
-            $output['country'] = (isset($countries[$record->country->isoCode])) ? $countries[$record->country->isoCode] : $locale->getText('locale.country.unknown');
+            $record = $geodb->get($ip);
+            $output['countryCode'] = $record['country']['iso_code'];
+            $output['country'] = (isset($countries[$record['country']['iso_code']])) ? $countries[$record['country']['iso_code']] : $locale->getText('locale.country.unknown');
             //$output['countryTimeZone'] = DateTimeZone::listIdentifiers(DateTimeZone::PER_COUNTRY, $record->country->isoCode);
-            $output['continent'] = (isset($continents[$record->continent->code])) ? $continents[$record->continent->code] : $locale->getText('locale.country.unknown');
-            $output['continentCode'] = $record->continent->code;
-            $output['eu'] = (\in_array($record->country->isoCode, $eu)) ? true : false;
+            $output['continent'] = (isset($continents[$record['continent']['code']])) ? $continents[$record['continent']['code']] : $locale->getText('locale.country.unknown');
+            $output['continentCode'] = $record['continent']['code'];
+            $output['eu'] = (\in_array($record['country']['iso_code'], $eu)) ? true : false;
 
             foreach ($currencies as $code => $element) {
-                if (isset($element['locations']) && isset($element['code']) && \in_array($record->country->isoCode, $element['locations'])) {
+                if (isset($element['locations']) && isset($element['code']) && \in_array($record['country']['iso_code'], $element['locations'])) {
                     $currency = $element['code'];
                 }
             }