<!--
Thank you for sending the PR! We appreciate you spending the time to work on these changes.

Help us understand your motivation by explaining why you decided to make this change.

You can learn more about contributing to appwrite here: https://github.com/appwrite/appwrite/blob/master/CONTRIBUTING.md

Happy contributing!

-->

## What does this PR do?

Replaces current geoip2 mmdb reader with [maxmind reader](https://github.com/maxmind/MaxMind-DB-Reader-php) along with [native C extension](https://github.com/maxmind/libmaxminddb) which is noticeably faster.

## Test Plan

Compatible with previous test cases

## Related PRs and Issues

#474 

### Have you read the [Contributing Guidelines on issues](https://github.com/appwrite/appwrite/blob/master/CONTRIBUTING.md)?

Yes

## Doubtful Regarding :

I had to add a block from lines 142 to 153 in Dockerfile (which is already repeated at line 22 to 33 and 107 to 127) otherwise I was getting error no autoconf, no compiler found, etc, anything wrong here?

In line 46 Dockerfile, according to libmaxmindb docs, `ldconfig` should be run but in alpine `ldconfig` needs to be run with config directory(https://stackoverflow.com/questions/36990951/ldconfig-seems-no-functional-under-alpine-3-3), is that correct path on that line?

In line 164 Dockerfile, is that path correct to load extension for ini file?

I am not sure if any changes are needed in `docker-compose.yml`, `docker-compose.nginx.yml` and `Dockerfile.nginx`