@eldadfux 2020-10-23
Hey @hack3r-0m great work, this looks like a very promising start!

As for your question no need to make any changes in `docker-compose.yml`, `docker-compose.nginx.yml` and `Dockerfile.nginx`.

About the duplicated code part. We use a multistage Docker build, the first stage install all the composer dependencies. The second stage install all the PHP extensions and compile them. In the final stage we copy the vendor directory from the first stage and the `.so` files from the second stage and add them to the correct php.ini path.

It would be much easier for you to take a look at how we install the phpredis extension and replicate that to make the maxmind one work.

If you need more advice, let me know on Discord.