@eldadfux 2020-10-23
This is the section where we download and compile extensions:
https://github.com/appwrite/appwrite/blob/198b6e02b27e73868952f2a95cd0841a5c435884/Dockerfile#L48

This is the section where we copy the compiled .so files:
https://github.com/appwrite/appwrite/blob/198b6e02b27e73868952f2a95cd0841a5c435884/Dockerfile#L132

This is the section where we enable the extenstions in php.ini:
https://github.com/appwrite/appwrite/blob/198b6e02b27e73868952f2a95cd0841a5c435884/Dockerfile#L201

This way our final image is much smaller and don't include build tools that are unnecessary and considered not secure for production work.