@hack3r-0m 2020-10-23
alright! and PHP test cases which are failing will be solved by just making method change to api/account.php as it also had some geoip2 usage