@eldadfux 2020-10-23
Exactly... this also reminds me that api/users.php also has some uses of the geodb