@eldadfux 2020-10-23
When you have some time, please also submit the results of the benchmarks on this PR. It's important that we'll have records of the change effect, and I will also want to share (and credit 😉) the huge performance improvement you managed to have with our community.