@hack3r-0m 2020-10-24
## BENCHMARKS

OLD: geoip2 reader without any extension
AVG(5 runs): 36.2 seconds for 100k searches (running after breaks, less cache) and 34.7 seconds (running frequently, more cache)

NEW: maxmind PHP mmdb reader with native C extension
AVG(5 runs): 29.1 seconds for 100k searches (running after breaks, less cache) and 28.9 seconds (running frequently, more cache)